requirements.txt
tensorflow
opencv-python
numpy
matplotlib
scipy
pillow
streamlit
uvicorn
fastapi
python-multipart
requests